feat: SupplierArrivalLineDetailFixedItems
